Modularize graphql schema
