+++ backend/graphql/schemas/category.js
@@ -1,0 +1,6 @@
+export const CategorySchema = `
+    type Category {
+        _id: ID!
+        name: String!
+    }
+`;
