+++ backend/graphql/schemas/index.js
@@ -1,184 +1,23 @@
 import { buildSchema } from 'graphql';
+import { UserSchema } from './user.js';
+import { CategorySchema } from './category.js';
+import { ProductSchema } from './product.js';
+import { OrderSchema } from './order.js';
+import { QuestionSchema } from './question.js';
+import { ProductFilterSchema } from './productFilter.js';
 
 export default buildSchema(`
+    ${UserSchema}
 
-    type User {
-        _id: ID!
-        name: String!
-        phoneNo: Int!
-        email: String!
-        password: String!
-        isAdmin: Boolean!
-        token: String
-    }
+    ${CategorySchema}
 
-    type Product {
-        _id: ID!
-        name: String!,
-        price: Float!,
-        user: User!,
-        image: String!,
-        brand: String!,
-        category: ID!,
-        countInStock: Int!,
-        numReviews: Int,
-        description: String!
-    }
+    ${ProductSchema}
 
-    type ProductResponse {
-        _id: ID!
-        name: String!,
-        price: Float!,
-        user: User!,
-        image: String!,
-        brand: ID!,
-        category: ID!,
-        countInStock: Int!,
-        numReviews: Int,
-        description: String!
-    }
-    
-    type OrderItems {
-        name: String!
-        qty: Float!
-        image: String!
-        price: Float!
-        product: Product!
-    }
+    ${OrderSchema}
 
-    type ShippingAddress {
-        address: String!
-        city: String!
-        postalCode: String!
-        country: String!
-    }
+    ${QuestionSchema}
 
-    type PaymentResult {
-        id: String!
-        status: String!
-        update_time: String!
-        email_address: String!
-    }
-
-    type Order {
-        _id:             ID!
-        user:            User!
-        orderItems:      [OrderItems!]!
-        shippingAddress: ShippingAddress!
-        paymentMethod:   String!
-        paymentResult:   PaymentResult
-        taxPrice:        Float!
-        shippingPrice:   Float!
-        totalPrice:      Float!
-        isPaid:          Boolean!
-        paidAt:          String
-        isDelivered:     Boolean!
-        deliveredAt:     String
-    }
-    
-    type Question {
-        _id: ID
-        msg: String
-        info: String
-        level: String
-        index: String
-    }
-
-    type Category {
-        _id: ID!
-        name: String!
-    }
-    
-    type Response {
-        msg: String!
-    }
-
-    input UserInput {
-        name: String!
-        phoneNo: Int!
-        email: String!
-        password: String!
-        isAdmin: Boolean
-    }
-
-    input UpdateUserInput {
-        name: String
-        phoneNo: Int!
-        email: String
-        password: String
-        isAdmin: Boolean
-    }
-
-    input ProductInput {
-        name: String!,
-        price: Float!,
-        user: ID!,
-        image: String!,
-        brand: String!,
-        category: String!,
-        countInStock: Int!,
-        numReviews: Int,
-        description: String!
-    }
-            
-    input OrderItemsInput {
-        name: String!
-        qty: Float!
-        image: String!
-        price: Float!
-        product: ID!
-    }
-
-    input ShippingAddressInput {
-        address: String!
-        city: String!
-        postalCode: String!
-        country: String!
-    }
-
-    input PaymentResultInput {
-        id: String!
-        status: String!
-        update_time: String!
-        email_address: String!
-    }
-
-    input updateProduct {
-        name: String,
-        price: Float,
-        image: String,
-        brand: String,
-        category: String,
-        countInStock: Int,
-        description: String,
-    }
-    
-    input OrderInput {
-        orderItems:      [OrderItemsInput!]!
-        shippingAddress: ShippingAddressInput!
-        paymentMethod:   String!
-        paymentResult:   PaymentResultInput
-        taxPrice:        Float!
-        shippingPrice:   Float!
-        totalPrice:      Float!
-        isPaid:          Boolean!
-        paidAt:          String
-        isDelivered:     Boolean!
-        deliveredAt:     String
-    }
-
-    input QuestionInput {
-        msg: String
-        info: String
-        level: String
-        index: String
-    }
-
-    input FilterInput {
-        brand: String
-        price: String
-        rating: String
-    }
+    ${ProductFilterSchema}
 
     type rootQuery {
         orders: [Order!]!
