+++ backend/graphql/schemas/order.js
@@ -1,0 +1,76 @@
+export const OrderSchema = `
+    type OrderItems {
+        name: String!
+        qty: Float!
+        image: String!
+        price: Float!
+        product: Product!
+    }
+
+    type ShippingAddress {
+        address: String!
+        city: String!
+        postalCode: String!
+        country: String!
+    }
+
+    type PaymentResult {
+        id: String!
+        status: String!
+        update_time: String!
+        email_address: String!
+    }
+
+    type Order {
+        _id:             ID!
+        user:            User!
+        orderItems:      [OrderItems!]!
+        shippingAddress: ShippingAddress!
+        paymentMethod:   String!
+        paymentResult:   PaymentResult
+        taxPrice:        Float!
+        shippingPrice:   Float!
+        totalPrice:      Float!
+        isPaid:          Boolean!
+        paidAt:          String
+        isDelivered:     Boolean!
+        deliveredAt:     String
+    }
+
+            
+    input OrderItemsInput {
+        name: String!
+        qty: Float!
+        image: String!
+        price: Float!
+        product: ID!
+    }
+
+    input ShippingAddressInput {
+        address: String!
+        city: String!
+        postalCode: String!
+        country: String!
+    }
+
+    input PaymentResultInput {
+        id: String!
+        status: String!
+        update_time: String!
+        email_address: String!
+    }
+        
+    input OrderInput {
+        orderItems:      [OrderItemsInput!]!
+        shippingAddress: ShippingAddressInput!
+        paymentMethod:   String!
+        paymentResult:   PaymentResultInput
+        taxPrice:        Float!
+        shippingPrice:   Float!
+        totalPrice:      Float!
+        isPaid:          Boolean!
+        paidAt:          String
+        isDelivered:     Boolean!
+        deliveredAt:     String
+    }
+`;
