+++ backend/graphql/schemas/productFilter.js
@@ -1,0 +1,7 @@
+export const ProductFilterSchema = `
+    input FilterInput {
+        brand: String
+        price: String
+        rating: String
+    }
+`;
