+++ backend/graphql/schemas/question.js
@@ -1,0 +1,20 @@
+export const QuestionSchema = `
+    type Question {
+        _id: ID
+        msg: String
+        info: String
+        level: String
+        index: String
+    }
+
+    type Response {
+        msg: String!
+    }
+
+    input QuestionInput {
+        msg: String
+        info: String
+        level: String
+        index: String
+    }
+`;
