+++ backend/models/subcategoryModel.js
@@ -1,0 +1,21 @@
+import mongoose from 'mongoose';
+
+const subCategorySchema = new mongoose.Schema(
+  {
+    name: {
+      type: String,
+      required: true,
+    },
+    category: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Category',
+    },
+  },
+  {
+    timestamps: true,
+  }
+);
+
+const Category = mongoose.model('SubCategory', subCategorySchema);
+
+export default Category;
